+++ packages/botonic-plugin-flow-builder/src/content-fields/flow-button.tsx
@@ -1,20 +1,17 @@
-import { Button, Reply } from '@botonic/react'
+import { Button, Reply, Webview } from '@botonic/react'
 import React from 'react'
 
 import { FlowBuilderApi } from '../api'
 import { SOURCE_INFO_SEPARATOR } from '../constants'
 import { ContentFieldsBase } from './content-fields-base'
-import {
-  HtButton,
-  HtButtonStyle,
-  HtPayloadNode,
-  HtUrlNode,
-} from './hubtype-fields'
+import { HtButton, HtButtonStyle, HtUrlNode } from './hubtype-fields'
 
 export class FlowButton extends ContentFieldsBase {
   public text = ''
   public url?: string
   public payload?: string
+  public target?: string
+  public webview?: Webview
 
   static fromHubtypeCMS(
     cmsButton: HtButton,
@@ -54,6 +51,8 @@ export class FlowButton extends ContentFieldsBase {
         key={this.id}
         url={this.url}
         payload={this.getButtonPayload(buttonIndex)}
+        target={this.target}
+        webview={this.webview}
       >
         {this.text}
       </Button>